+++ src/loofah/proactor/proact_acceptor.cpp
@@ -87,12 +87,12 @@ socket_t ProactAcceptorBase::get_socket() const
     return _listener_socket;
 }
 
-void ProactAcceptorBase::handle_read_completed(int cb)
+void ProactAcceptorBase::handle_read_completed(ssize_t cb)
 {
     // Dummy for an acceptor
 }
 
-void ProactAcceptorBase::handle_write_completed(int cb)
+void ProactAcceptorBase::handle_write_completed(ssize_t cb)
 {
     // Dummy for an acceptor
 }
