+++ src/loofah/proactor/proact_acceptor.h
@@ -24,9 +24,9 @@ public:
 
     virtual socket_t get_socket() const override;
 
-    virtual void handle_read_completed(int cb) final override;
+    virtual void handle_read_completed(ssize_t cb) final override;
 
-    virtual void handle_write_completed(int cb) final override;
+    virtual void handle_write_completed(ssize_t cb) final override;
 
 private:
     socket_t _listener_socket = LOOFAH_INVALID_SOCKET_FD;
